requirements.txt
maltego-trx
py3-validate-email==0.2.12
